fix assert name

+++ packages/backend-for-frontend/src/routers/clientRouter.ts
@@ -14,7 +14,7 @@ import { makeApiProblem } from "../model/errors.js";
 import { PagoPAInteropBeClients } from "../clients/clientsProvider.js";
 import { fromBffAppContext } from "../utilities/context.js";
 import { getClientUsersErrorMapper } from "../utilities/errorMappers.js";
-import { assertEndpointIsCallable } from "../services/validators.js";
+import { assertAdminClientFeatureEnabled } from "../services/validators.js";
 
 const clientRouter = (
   ctx: ZodiosContext,
@@ -349,7 +349,7 @@ const clientRouter = (
     })
 
     .delete("/clients/:clientId/admin/:adminId", async (req, res) => {
-      assertEndpointIsCallable(res);
+      assertAdminClientFeatureEnabled(res);
       const ctx = fromBffAppContext(req.ctx, req.headers);
       try {
         await clientService.removeClientAdmin(
